* Removed quartz dependency completely to server. It's just not needed on client * Better management of GUI callback connections (separated to GUIRegistry). This fixes a number of timing related issues such as resources appearing twice. * Changed log output of prompt parameters to debug (to hide passwords) * Connect/Disconnect/Delete buttons now sit next to the preset selection * Completed stay connected/connect on startup

+++ client-service/src/main/java/com/hypersocket/client/service/ConnectionServiceImpl.java
@@ -66,8 +66,10 @@ public class ConnectionServiceImpl implements ConnectionService {
 		crit.add(Restrictions.eq("id", con.getId()));
 		
 		Connection toDelete = (Connection) crit.uniqueResult();
-		session.delete(toDelete);
-		session.flush();
+		if(toDelete != null) {
+			session.delete(toDelete);
+			session.flush();
+		}
 		trans.commit();
 		
 	}
